+++ pandas/core/arrays/arrow/__init__.py
@@ -1,4 +1,3 @@
 from pandas.core.arrays.arrow.array import ArrowExtensionArray
-from pandas.core.arrays.arrow.dtype import ArrowDtype
 
-__all__ = ["ArrowDtype", "ArrowExtensionArray"]
+__all__ = ["ArrowExtensionArray"]
